Dockerhub, search namespaces and repos impl, not tested yet

+++ src/connectors/dockerhub_cservice.rs
@@ -124,7 +124,7 @@ impl RedisCache {
 
         let mut conn = self.connection.lock().await;
         let (): () = conn
-            .set_ex(key, payload, ttl_secs as usize)
+            .set_ex(key, payload, ttl_secs)
             .await
             .map_err(|err| ConnectorError::ServiceUnavailable(format!("Redis SET failed: {}", err)))?;
         Ok(())
